Some fixes 2/2 - modules refactored
Service:
- unary service:  server module
- fs2 service: fs2 module
- monix service: monix module

Example:
- health-client: all clients
- health-server-fs2: fs2 server
- health-server-monix: monix server
