--- modules/examples/health-check-monix/health-client/src/main/resources/logback.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<configuration scan="false" debug="false">
-    <appender name="stdout" class="ch.qos.logback.core.ConsoleAppender">
-        <encoder>
-            <pattern>%date{HH:mm:ss} %-5level %class - %msg%n</pattern>
-        </encoder>
-    </appender>
-    <root level="INFO">
-        <appender-ref ref="stdout" />
-    </root>
-</configuration>
