+++ modules/server/src/main/scala/higherkindness/mu/rpc/healthcheck/ServerStatus.scala
